Tarea9

+++ src/main/java/uce/edu/web/api/repository/modelo/Profesor.java
@@ -1,5 +1,6 @@
 package uce.edu.web.api.repository.modelo;
 
+import java.math.BigDecimal;
 import jakarta.persistence.*;
 
 @Entity
@@ -18,48 +19,62 @@ public class Profesor {
     @Column(name = "prof_tipo_contrato")
     private String tipoContrato;
     @Column(name = "prof_salario")
-    private float salario;
+    private BigDecimal salario;
     @Column(name = "prof_email")
     private String email;
+
     public Integer getId() {
         return id;
     }
+
     public void setId(Integer id) {
         this.id = id;
     }
+
     public String getNombre() {
         return nombre;
     }
+
     public void setNombre(String nombre) {
         this.nombre = nombre;
     }
+
     public String getApellido() {
         return apellido;
     }
+
     public void setApellido(String apellido) {
         this.apellido = apellido;
     }
+
     public String getAsignatura() {
         return asignatura;
     }
+
     public void setAsignatura(String asignatura) {
         this.asignatura = asignatura;
     }
+
     public String getTipoContrato() {
         return tipoContrato;
     }
+
     public void setTipoContrato(String tipoContrato) {
         this.tipoContrato = tipoContrato;
     }
-    public float getSalario() {
+
+    public BigDecimal getSalario() {
         return salario;
     }
-    public void setSalario(float salario) {
+
+    public void setSalario(BigDecimal salario) {
         this.salario = salario;
     }
+
     public String getEmail() {
         return email;
     }
+
     public void setEmail(String email) {
         this.email = email;
     }
